fix: revision on trade supplier fuel report v2
