--- IBSWeb/Areas/Filpride/Views/AccountsPayableReport/TradeSupplierCommissioneeReport.cshtml
@@ -1,59 +1,0 @@
-@model ViewModelBook
-
-@{
-    ViewData["Title"] = "Accounts Payable Report - Trade Supplier Commissionee Report";
-}
-
-<div class="card shadow border-0 mt-4">
-    <div class="card-header bg-secondary bg-gradient ml-0 py-3">
-        <div class="row">
-            <div class="col-12 text-center">
-                <h2 class="text-white py-2">@ViewData["Title"]</h2>
-            </div>
-        </div>
-    </div>
-    <div class="card-body p-4">
-        <form id="dynamicForm" method="post" target="_blank">
-            <div class="row">
-                <div class="mb-3">
-                    <label asp-for="DateFrom" class="control-label"></label>
-                    <input type="date" asp-for="DateFrom" class="form-control" value="@DateTimeHelper.GetCurrentPhilippineTime().ToString("yyyy-MM-dd")" />
-                    <span asp-validation-for="DateFrom" class="text-danger"></span>
-                </div>
-                <div class="mb-3">
-                    <label asp-for="DateTo" class="control-label"></label>
-                    <input type="date" asp-for="DateTo" class="form-control" value="@DateTimeHelper.GetCurrentPhilippineTime().ToString("yyyy-MM-dd")" />
-                    <span asp-validation-for="DateTo" class="text-danger"></span>
-                </div>
-                <div class="mb-3">
-                    <label asp-for="StatusFilter" class="control-label"></label>
-                    <select asp-for="StatusFilter" class="form-select">
-                        <option value="ValidOnly">Valid Only</option>
-                        <option value="All">All (Include Voided)</option>
-                        <option value="InvalidOnly">Voided Only</option>
-                    </select>
-                    <span asp-validation-for="StatusFilter" class="text-danger"></span>
-                </div>
-                <div class="row pt-2">
-                    <div class="col-6 col-md-3">
-                        <button type="button" class="btn btn-primary form-control" onclick="submitForm('GenerateTradeSupplierCommissioneeReportExcelFile')"><i class="bi bi-file-text"></i> Generate as Excel file</button>
-                    </div>
-                </div>
-            </div>
-        </form>
-    </div>
-</div>
-
-@section Scripts {
-    @{
-        await Html.RenderPartialAsync("_ValidationScriptsPartial");
-    }
-    <script>
-        function submitForm(action) {
-            var form = document.getElementById('dynamicForm');
-            form.action = action;
-            form.submit();
-        }
-    </script>
-}
-
