+++ IBSWeb/Areas/Filpride/Views/SubsidiaryLedgerReport/TradeCommissioneeReport.cshtml
@@ -1,0 +1,53 @@
+@model ViewModelBook
+
+@{
+    ViewData["Title"] = "Accounts Payable Report - Trade Supplier Commissionee Report";
+}
+
+<div class="card shadow border-0 mt-4">
+    <div class="card-header bg-secondary bg-gradient ml-0 py-3">
+        <div class="row">
+            <div class="col-12 text-center">
+                <h2 class="text-white py-2">@ViewData["Title"]</h2>
+            </div>
+        </div>
+    </div>
+    <div class="card-body p-4">
+        <form id="dynamicForm" method="post" target="_blank">
+            <div class="row">
+                <div class="mb-4 row align-items-center">
+                    <label for="monthDate" class="col-sm-4 col-form-label text-end fw-semibold">
+                        As Of (Month and Year)
+                    </label>
+                    <div class="col-sm-6">
+                        <input type="month"
+                               id="monthDate"
+                               name="monthDate"
+                               class="form-control"
+                               value="@DateTimeHelper.GetCurrentPhilippineTime().ToString("yyyy-MM")"
+                               required />
+                    </div>
+                </div>
+                <div class="row pt-2">
+                    <div class="col-6 col-md-3">
+                        <button type="button" class="btn btn-primary form-control" onclick="submitForm('GenerateTradeCommissioneeReportExcelFile')"><i class="bi bi-file-text"></i> Generate as Excel file</button>
+                    </div>
+                </div>
+            </div>
+        </form>
+    </div>
+</div>
+
+@section Scripts {
+    @{
+        await Html.RenderPartialAsync("_ValidationScriptsPartial");
+    }
+    <script>
+        function submitForm(action) {
+            var form = document.getElementById('dynamicForm');
+            form.action = action;
+            form.submit();
+        }
+    </script>
+}
+
