+++ IBSWeb/Areas/Filpride/Views/SubsidiaryLedgerReport/TradeFuelReport.cshtml
@@ -15,7 +15,6 @@
     <div class="card-body p-4">
         <form id="dynamicForm" method="post" target="_blank">
             <div class="row">
-
                 <div class="mb-4 row align-items-center">
                     <label for="monthDate" class="col-sm-4 col-form-label text-end fw-semibold">
                         As Of (Month and Year)
@@ -29,29 +28,9 @@
                                required />
                     </div>
                 </div>
-
-                <div class="mb-4 row align-items-center">
-                    <label asp-for="StatusFilter"
-                           class="col-sm-4 col-form-label text-end fw-semibold">
-                    </label>
-                    <div class="col-sm-6">
-                        <select asp-for="StatusFilter" class="form-select">
-                            <option value="ValidOnly">Valid Only</option>
-                            <option value="All">All (Include Voided)</option>
-                            <option value="InvalidOnly">Voided Only</option>
-                        </select>
-                        <span asp-validation-for="StatusFilter" class="text-danger"></span>
-                    </div>
-                </div>
-
                 <div class="row pt-2">
                     <div class="offset-sm-4 col-sm-6">
-                        <button type="button"
-                                class="btn btn-primary w-100"
-                                onclick="submitForm('GenerateTradeSupplierFuelReportExcelFile')">
-                            <i class="bi bi-file-text"></i>
-                            Generate as Excel file
-                        </button>
+                        <button type="button" class="btn btn-primary w-100" onclick="submitForm('GenerateTradeFuelReportExcelFile')"><i class="bi bi-file-text"></i> Generate as Excel file</button>
                     </div>
                 </div>
